\name{Chess}
\docType{data}
\alias{Chess}

\encoding{utf8}

\title{Chess dataset}

\description{
The Chess data describes chess games of certain kind described by attributes 
representing the positions of pieces on the board. More info: http://fimi.ua.ac.be/data/
dimension: 3196 objects, 76 attributes, number of values in the scale of grades: 2
  Downloaded from: http://datasets.inf.upol.cz/
}

\usage{data(Chess)}

\format{Binary matrix, name Chess}

\references{
Lichman, M: UCI Machine Learning Repository http://archive.ics.uci.edu/ml. Irvine, CA: 
University of California, School of Information and Computer Science, 2013
}

\keyword{datasets}
\keyword{Chess}