\name{DBLP}
\docType{data}
\alias{DBLP}

\encoding{utf8}

\title{DBLP dataset}

\description{
The DBLP dataset contains records of in which of the 19 conferences the 6980 authors had published.
 The dataset is collected from the DBLP database .
  dimension: 6980 objects, 19 attributes, number of values in the scale of grades: 2
  Downloaded from: http://datasets.inf.upol.cz/
}

\usage{data(DBLP)}

\format{Binary matrix, name DBLP}

\references{
Miettinen, P.: Matrix Decomposition Methods for Data Mining: 
Computational Complexity and Algorithms, PhD thesis, 
University of Helsinki, 2009
}

\keyword{datasets}
\keyword{DBLP}